refactor(core): extract WordPiece Unicode category tables
Move BertNormalizer/pre-tokenizer category helpers and the
control/punctuation range tables into wordpiece/unicode.rs.
